Changed character Classes
