+++ src/game/characters/archer/Saggitarius.java
@@ -1,0 +1,11 @@
+package game.characters.archer;
+
+import game.Player;
+import game.characters.Archer;
+import game.characters.Category;
+
+public class Saggitarius extends Archer {
+    public Saggitarius(Player owner) {
+        super(owner, "Saggitarius", Category.Mystics, 230, 18, 7, 12, 17);
+    }
+}
