+++ src/game/characters/archer/Zing.java
@@ -1,0 +1,11 @@
+package game.characters.archer;
+
+import game.Player;
+import game.characters.Archer;
+import game.characters.Category;
+
+public class Zing extends Archer {
+    public Zing(Player owner) {
+        super(owner, "Zing", Category.Sunchildren, 200, 16, 9, 11, 14);
+    }
+}
