+++ src/game/characters/healer/Lightbringer.java
@@ -1,0 +1,11 @@
+package game.characters.healer;
+
+import game.Player;
+import game.characters.Category;
+import game.characters.Healer;
+
+public class Lightbringer extends Healer {
+    public Lightbringer(Player owner) {
+        super(owner, "Lightbringer", Category.Sunchildren, 260, 17, 15, 19, 12);
+    }
+}
