+++ src/game/characters/knight/Swiftblade.java
@@ -1,0 +1,11 @@
+package game.characters.knight;
+
+import game.Player;
+import game.characters.Category;
+import game.characters.Knight;
+
+public class Swiftblade extends Knight {
+    public Swiftblade(Player owner) {
+        super(owner, "Swiftblade", Category.Marshlanders, 250, 18, 20, 17, 13);
+    }
+}
