+++ src/game/characters/knight/Templar.java
@@ -1,0 +1,11 @@
+package game.characters.knight;
+
+import game.Player;
+import game.characters.Category;
+import game.characters.Knight;
+
+public class Templar extends Knight {
+    public Templar(Player owner) {
+        super(owner, "Templar", Category.Sunchildren, 155, 14, 16, 12, 12);
+    }
+}
