+++ src/game/characters/knight/Zoro.java
@@ -1,0 +1,11 @@
+package game.characters.knight;
+
+import game.Player;
+import game.characters.Category;
+import game.characters.Knight;
+
+public class Zoro extends Knight {
+    public Zoro(Player owner) {
+        super(owner, "Zoro", Category.Highlanders, 180, 17, 16, 13, 14);
+    }
+}
